Deleted second mail from repository to same author, Dave Hansen <dave@sr71.net>. Merged another three patches. Look previous logs.
Now you can start to see how the TileSet gets used.  In this
case, we're replacing the loadAllTiles() iteration with a
TileSet instead.  Saves something like 50% of the lines of
code in that function.

In addition to using a TileSet, this also uses the Tile
class's pixelPos() function.

imgToTile is here to make the imageUpdate()
function simpler.
This also adds a nice little message about how many
tiles we've been downloading.

Just another two to be complete.

git-svn-id: http://svn.openstreetmap.org/applications/editors/josm/plugins@16161 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4